New effects implemented

+++ app/src/main/java/com/shebut_dev/photoeditor/core/effects/EffectNegative.java
@@ -7,7 +7,7 @@ import com.shebut_dev.photoeditor.core.BaseEffect;
 //Реализаци эффекта "Негатив"
 public class EffectNegative implements BaseEffect {
     @Override
-    public Bitmap apply(Bitmap sourceImage, int intensity) {
+    public Bitmap apply(Bitmap sourceImage, float intensity) {
         return sourceImage;
     }
 }
